+++ packages/extension/providers/connection-explorer.ts
@@ -1,5 +1,3 @@
-import fs from 'fs';
-import path from 'path';
 import {
   Event,
   EventEmitter,
@@ -14,7 +12,7 @@ import {
   SidebarView,
 } from './sidebar-provider/sidebar-tree-items';
 import { ConnectionCredentials, SerializedConnection } from '@sqltools/core/interface';
-
+import { Logger } from '../api';
 export type SidebarDatabaseItem = SidebarConnection | SidebarTable | SidebarColumn | SidebarView;
 
 export class ConnectionExplorer implements TreeDataProvider<SidebarDatabaseItem> {
@@ -22,6 +20,7 @@ export class ConnectionExplorer implements TreeDataProvider<SidebarDatabaseItem>
   public readonly onDidChangeTreeData: Event<SidebarDatabaseItem | undefined> =
   this.onDidChange.event;
   private tree: { [database: string]: SidebarConnection} = {};
+  constructor(private logger: Logger) { }
   public fireUpdate(): void {
     this.onDidChange.fire();
   }
@@ -73,11 +72,20 @@ export class ConnectionExplorer implements TreeDataProvider<SidebarDatabaseItem>
 
     if (this.tree[treeKey]) this.tree[treeKey].reset();
 
+    if (!tables && !columns) {
+      this.tree[treeKey].deactivate();
+      return this.refresh();
+    }
+
     tables.sort((a, b) => a.name.localeCompare(b.name)).forEach((item) => {
       if (!this.tree[treeKey]) return;
       this.tree[treeKey].addItem(item);
     });
     columns.sort((a, b) => a.columnName.localeCompare(b.columnName)).forEach((column) => {
+      if (!column || !this.tree[treeKey] || this.tree[treeKey].views) {
+        this.logger.warn('didnt find a place in three for column', column);
+        return;
+      }
       const key = this.tree[treeKey].views.items[column.tableName] ? 'views' : 'tables';
       this.tree[treeKey][key].items[column.tableName].addItem(new SidebarColumn(column, conn));
     });