+++ packages/language-server/store/index.ts
@@ -4,6 +4,7 @@ import * as actions from './actions';
 import { DatabaseInterface } from '@sqltools/core/interface';
 
 interface State {
+  lastUsedId: string;
   activeConnections: { [id: string]: Connection };
   queryResults: {
     [connection_id: string]: {
@@ -13,6 +14,7 @@ interface State {
 }
 
 const initialState: State = {
+  lastUsedId: undefined,
   activeConnections: {},
   queryResults: {}
 }
@@ -20,14 +22,18 @@ const initialState: State = {
 const onConnect = (state: State, conn: Connection) => {
   const activeConnections = Object.assign({}, state.activeConnections, { [conn.getId()]: conn });
   const queryResults = Object.assign({}, state.queryResults, { [conn.getId()]: {} });
-  return Object.assign({}, state, { activeConnections, queryResults });
+
+  return Object.assign({}, state, { activeConnections, queryResults, lastUsedId: conn.getId() });
 };
 
 const onDisconnect = (state: State, conn: Connection) => {
-  const { activeConnections, queryResults } = state;
+  const { activeConnections, queryResults, lastUsedId } = state;
   delete activeConnections[conn.getId()];
   delete queryResults[conn.getId()];
-  return Object.assign({}, state, { activeConnections, queryResults });
+
+  const newLastUsedId = lastUsedId === conn.getId() ? undefined : lastUsedId;
+
+  return Object.assign({}, state, { activeConnections, queryResults, lastUsedId: newLastUsedId });
 };
 
 const onQuerySuccess = (state: State, { conn, query, results }: { conn: Connection, query: string, results: DatabaseInterface.QueryResults }) => {