+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 
 ## v0.16
 
+### v0.16.4
+
+- Fixes disconnect not working. #122.
+- Auto connecting if have just one connection
+- 'Add new Server' changed to 'Add new connection'
+- Fixes show records when not connected
+
 ### v0.16.3
 
 - Show records and describe tables when using multiple connections. #119